abstracted the scp function to a different file
